Read Description
Removed zucc,markdown,elonmusk,facepalm and despacito due to useless commands and markdown used a website that has died a long time ago. Also removed to maestr0 due to me making a credits commands for this probs.  Added spam command, but please don't use it on not-private channels. You'll get banned in the server and discord too. Use the spam feature to farm pokemon in a private channel on your own server.